Removed writers, use builder (again)
