+++ pyqrcode/__init__.py
@@ -43,7 +43,6 @@ Examples:
 from __future__ import absolute_import, division, print_function, with_statement, unicode_literals
 import pyqrcode.tables
 import pyqrcode.builder as builder
-from pyqrcode import writers
 try:  # pragma: no cover
     str = unicode  # Python 2
 except NameError:
@@ -566,7 +565,7 @@ class QRCode:
         left settable because such a wide quiet zone is unnecessary in many
         applications where the QR code is not being printed.
         """
-        return writers.write_xbm(self.code, scale, quiet_zone)
+        return builder._xbm(self.code, scale, quiet_zone)
 
     def svg(self, file, scale=1, module_color='#000', background=None,
             quiet_zone=4, xmldecl=True, svgns=True, title=None,
@@ -622,11 +621,11 @@ class QRCode:
             >>> code.svg('live-organ-transplants.svg', scale=4,
                          module_color='brown', background='0xFFFFFF')
         """
-        writers.write_svg(self.code, self.version, file, scale=scale,
-                          module_color=module_color, background=background,
-                          quiet_zone=quiet_zone, xmldecl=xmldecl, svgns=svgns,
-                          title=title, svgclass=svgclass, lineclass=lineclass,
-                          omithw=omithw, debug=debug)
+        builder._svg(self.code, self.version, file, scale=scale,
+                     module_color=module_color, background=background,
+                     quiet_zone=quiet_zone, xmldecl=xmldecl, svgns=svgns,
+                     title=title, svgclass=svgclass, lineclass=lineclass,
+                     omithw=omithw, debug=debug)
 
     def eps(self, file, scale=1, module_color=(0, 0, 0),
             background=None, quiet_zone=4):
@@ -660,8 +659,8 @@ class QRCode:
             >>> out = io.StringIO()
             >>> qr.eps(out, module_color=(.4, .4, .4))
         """
-        writers.write_eps(self.code, self.version, file, scale, module_color,
-                          background, quiet_zone)
+        builder._eps(self.code, self.version, file, scale, module_color,
+                     background, quiet_zone)
 
     def terminal(self, module_color='default', background='reverse',
                  quiet_zone=4):
@@ -708,8 +707,7 @@ class QRCode:
             >>> text = code.terminal()
             >>> print(text)
         """
-        return writers.write_terminal(self.code, module_color, background,
-                                      quiet_zone)
+        return builder._terminal(self.code, module_color, background, quiet_zone)
 
     def text(self, quiet_zone=4):
         """This method returns a string based representation of the QR code.
@@ -727,4 +725,4 @@ class QRCode:
             >>> text = code.text()
             >>> print(text)
         """
-        return writers.write_text(self.code, quiet_zone)
+        return builder._text(self.code, quiet_zone)
